function pointers

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,4 @@
 #include "function_pointers.h"
-#include <stddef.h>
 /**
  * array_iterator - executes a function given as a parameter
  * on each element of an array
@@ -9,7 +8,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned int i;
 
 	if (array[size] && action)
